% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{control}
\alias{control}
\title{Configure IPM projection settings}
\usage{
control(years = 100, delta_time = 1, store_every = 1, bin_width = 1)
}
\arguments{
\item{years}{Positive integer. Number of simulation timesteps. Default 100.}

\item{delta_time}{Positive numeric. Duration of each timestep in years. Default 1.}

\item{store_every}{Positive integer. Store stand state every N timesteps. Default 1.}

\item{bin_width}{Positive integer. Bin width for IPM kernel discretization. Default 1.}
}
\value{
An object of S3 class \code{"ipm_control"}.
}
\description{
Configure IPM projection settings
}
