% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supported_species.R
\name{supported_species}
\alias{supported_species}
\title{Return the tibble of species supported by forestIPM}
\usage{
supported_species()
}
\value{
A tibble with six columns: species_id, common_name, nom_commun,
growth_model, surv_model, recruit_model. One row per supported species.
}
\description{
Return the tibble of species supported by forestIPM
}
\examples{
sp <- supported_species()
head(sp)
}
