% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{parameters}
\alias{parameters}
\title{Resolve a single parameter realization from Bayesian posteriors}
\usage{
parameters(mod, draw = "random", seed = NULL)
}
\arguments{
\item{mod}{An \code{ipm_spModel} object.}

\item{draw}{Character \code{"mean"} or \code{"random"}, or a positive integer
1-1000 selecting a specific posterior draw index.}

\item{seed}{Integer or NULL. Random seed used only when \code{draw = "random"}.}
}
\value{
An object of S3 class \code{"ipm_parameters"}.
}
\description{
Resolve a single parameter realization from Bayesian posteriors
}
