% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{plot.ipm_projection}
\alias{plot.ipm_projection}
\title{Plot lambda trajectory from an IPM projection}
\usage{
\method{plot}{ipm_projection}(x, y, ...)
}
\arguments{
\item{x}{An \code{ipm_projection} object.}

\item{y}{Ignored.}

\item{...}{Additional arguments passed to \code{plot()}.}
}
\description{
Plot lambda trajectory from an IPM projection
}
