% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project}
\alias{project}
\title{Project population dynamics over time}
\usage{
project(mod, pars, stand, env, ctrl)
}
\arguments{
\item{mod}{An \code{ipm_spModel} object.}

\item{pars}{An \code{ipm_parameters} object.}

\item{stand}{An \code{ipm_stand} object. Initial stand state (t = 0).}

\item{env}{An \code{ipm_env} object. Climate drivers (may be time-varying).}

\item{ctrl}{An \code{ipm_control} object. Required.}
}
\value{
An object of S3 class \code{"ipm_projection"} with fields:
\code{$species}, \code{$years}, \code{$lambda}, \code{$stand_series}
(list of \code{ipm_dist_snapshot} objects, each holding the continuous
size distribution per species at that timestep), \code{$summary}.
}
\description{
Project population dynamics over time
}
